refactor: move ObjectType to a separate file
