+++ query-node/substrate-query-framework/cli/src/model/ObjectType.ts
@@ -1,0 +1,16 @@
+import { Field } from '.';
+
+/**
+ * Reperesent GraphQL object type
+ */
+export interface ObjectType {
+  name: string;
+  fields: Field[];
+  isEntity: boolean;
+  // imports for relations
+  imports: Set<string>;
+  // Description of the field will be shown in GrapqQL API
+  description?: string;
+  isInterface?: boolean;
+  interfaces?: ObjectType[]; //interface names
+}
