+++ query-node/substrate-query-framework/cli/src/model/index.ts
@@ -1,5 +1,7 @@
 import { Field } from './Field';
+import { Relation } from './Relation'
 import { FTSQuery } from './FTSQuery';
-import { WarthogModel, ObjectType } from './WarthogModel';
+import { WarthogModel } from './WarthogModel';
+import { ObjectType } from './ObjectType'
 
-export { FTSQuery, WarthogModel, ObjectType, Field };
+export { FTSQuery, WarthogModel, ObjectType, Field, Relation };
